Revert age uint32 -> uint256 (#4)

+++ src/mosm.sol
@@ -65,7 +65,7 @@ contract Mosm is LibNote {
     }
 
     uint128        val;
-    uint32  public age;
+    uint256  public age;
     bytes32 public constant wat = "ethusd"; // You want to change this every deploy
     uint256 public bar = 1;
 
@@ -96,7 +96,7 @@ contract Mosm is LibNote {
         return (val, val > 0);
     }
 
-    function recover(uint256 val_, uint32 age_, uint8 v, bytes32 r, bytes32 s) internal pure returns (address) {
+    function recover(uint256 val_, uint256 age_, uint8 v, bytes32 r, bytes32 s) internal pure returns (address) {
         return ecrecover(
             keccak256(abi.encodePacked("\x19Ethereum Signed Message:\n32", keccak256(abi.encodePacked(val_, age_, wat)))),
             v, r, s
@@ -105,14 +105,14 @@ contract Mosm is LibNote {
 
     // Median
     function poke(
-        uint256[] calldata val_, uint32[] calldata age_,
+        uint256[] calldata val_, uint256[] calldata age_,
         uint8[] calldata v, bytes32[] calldata r, bytes32[] calldata s) external
     {
         require(val_.length == bar, "Median/bar-too-low");
 
         uint256 bloom = 0;
         uint256 last = 0;
-        uint32 snooze = age;
+        uint256 snooze = age;
 
         for (uint i = 0; i < val_.length; i++) {
             // Validate the values were signed by an authorized oracle
@@ -131,7 +131,7 @@ contract Mosm is LibNote {
         }
 
         val = uint128(val_[val_.length >> 1]);
-        age = uint32(block.timestamp);
+        age = block.timestamp;
 
         emit LogMedianPrice(val, age);
     }
@@ -305,15 +305,13 @@ contract Mosm is LibNote {
 }
 
 //// Create new deploy: Uncomment the below and modify contract name + wat ////
-/*
 contract MosmETHUSD is Mosm {
     bytes32 public constant wat = "ETHUSD";
 
-    function recover(uint256 val_, uint32 age_, uint8 v, bytes32 r, bytes32 s) internal pure returns (address) {
+    function recover(uint256 val_, uint256 age_, uint8 v, bytes32 r, bytes32 s) internal pure returns (address) {
         return ecrecover(
             keccak256(abi.encodePacked("\x19Ethereum Signed Message:\n32", keccak256(abi.encodePacked(val_, age_, wat)))),
             v, r, s
         );
     }
 }
-*/
